fix(fo4): five GUI-verified render fixes (trunk, cutout, LOD dupes, specular, clouds)
User GUI pass on FO4 surfaced five defects; each root-caused against the
raw assets:

- Tree trunks missing (TreeElmFree01): FO4 stores WIND WEIGHT in vertex
  alpha (trunk = 0) with SLSF1_Vertex_Alpha clear, and non-visual data in
  vertex RGB with SLSF2_Vertex_Colors clear. The extractor multiplied that
  alpha into the cutout test, discarding the whole trunk. Vertex channels
  are now gated by the SLSF flags for BSLightingShaderProperty shapes
  (wind alpha neutralized to 255, RGB dropped when the flag is clear);
  the FNV/legacy advisory default is untouched (saloon rust shading
  regression-checked).
- Sign letters as full rectangles (SignSuperDuperMart01): the letters
  shape has NO NiAlphaProperty and its material Name is an absolute
  developer build path (C:\Projects\Fallout4\...\SDMart1Letters.BGSM).
  ResolveMaterial now normalizes the path like the texture resolvers, so
  the BGSM alpha cutout actually applies.
- Purple z-fighting texture duplicates: BSMeshLODTriShape partitions its
  triangle buffer into standalone LOD slices; we drew the WHOLE buffer —
  full detail plus its simplified LOD copies on top. Now renders only the
  first non-empty slice (TreeElmFree01 LOD0=1191; VineHanging05 authors
  everything in LOD2).
- Specular blowout scene-wide: the BC5 uniform mask (1.0) is reverted —
  the material's _s map (FO4 slot 6 / FO76 slot 8, R = per-texel mask) is
  now bound at TexIndices.z and flagged via vTextureState.z; without an
  _s map specular stays off, as before S4.
- Cloud layers racing in nonsense directions: FO4/FO76 QNAM/RNAM are one
  FLOAT per layer, not FNV's signed byte — the parser read float bytes as
  speeds. WeatherRecord.CloudSpeedsX/Y are now normalized floats (-1..1)
  with a game-keyed parse (unit-tested both forms).

Decoded-mesh DecoderVersion 20->21 (LOD slice, vertex-alpha, material
resolvability, and new SpecularMapTexturePath all change the payload).

Verified headless: tree renders with trunk + bark; sign sunburst/letters
cut out; mailbox keeps masked specular; FNV saloon unchanged. 70 tests
across weather/BGSM/extractor/atmosphere green.

+++ src/BethesdaMultitool/Core/Formats/Esm/Models/Records/World/WeatherRecord.cs
@@ -65,13 +65,15 @@ public record WeatherRecord
     public IReadOnlyList<WeatherCloudAlpha> CloudLayerAlphas { get; init; } = [];
 
     /// <summary>
-    ///     QNAM "X Cloud Speeds" (FNV; xEdit wbWeatherCloudSpeed): one byte per cloud layer — the per-layer
-    ///     U-axis scroll rate the engine accumulates in <c>Clouds::Update</c>. Empty when absent.
+    ///     QNAM "X Cloud Speeds": per-layer U-axis scroll rate the engine accumulates in
+    ///     <c>Clouds::Update</c>, normalized to −1‥1 (0 = still). Authored as one signed byte per layer
+    ///     on FNV/FO3/Skyrim and one float per layer on FO4/FO76 — the handler normalizes both forms so
+    ///     the renderer never re-interprets raw bytes. Empty when absent.
     /// </summary>
-    public IReadOnlyList<byte> CloudSpeedsX { get; init; } = [];
+    public IReadOnlyList<float> CloudSpeedsX { get; init; } = [];
 
-    /// <summary>RNAM "Y Cloud Speeds" (FNV): per-layer V-axis scroll rate. Empty when absent.</summary>
-    public IReadOnlyList<byte> CloudSpeedsY { get; init; } = [];
+    /// <summary>RNAM "Y Cloud Speeds": per-layer V-axis scroll rate, normalized like <see cref="CloudSpeedsX" />.</summary>
+    public IReadOnlyList<float> CloudSpeedsY { get; init; } = [];
 
     /// <summary>DATA block (wind speed, sun glare, precipitation timing, flags, lightning color).</summary>
     public WeatherData? Data { get; init; }
